fix acc calculation bug

+++ train_DHN/utils.py
@@ -503,12 +503,11 @@ def eval_acc(score, target, weight, th=0.5):
                 predicted[b, h, int(indice)] = 1.0
         num_positive = float(target[b, :, :].sum())
         num_negative = float(target.size(1)*target.size(2) - num_positive)
-        num_tp = float(((predicted[b, :, :] == target[b, :, :]) + (target[b, :, :] == 1.0)).eq(2).sum())
-        num_tn = float(((predicted[b, :, :] == target[b, :, :]) + (target[b, :, :] == 0.0)).eq(2).sum())
+        num_tp = float(((predicted[b, :, :] == target[b, :, :]).float() + (target[b, :, :] == 1.0).float()).eq(2).sum())
+        num_tn = float(((predicted[b, :, :] == target[b, :, :]).float() + (target[b, :, :] == 0.0).float()).eq(2).sum())
 
         acc.append(1.0*(num_tp * float(weight[b, 1, 0, 0]) + num_tn * float(weight[b, 0, 0, 0]))/
                    (num_positive * float(weight[b, 1, 0, 0]) + num_negative * float(weight[b, 0, 0, 0])))
-
     return predicted, np.mean(np.array(acc))
